+++ R/metadata.R
@@ -1,0 +1,11 @@
+#' Vector of REDCap metadata headers
+#'
+#'
+#' @format Vector of length 18 with REDCap metadata headers:
+#' \describe{
+#'   \item{metadata_names}{characterstrings}
+#' }
+#' @seealso \url{https://www.project-redcap.org/}
+#' @usage data(metadata_names)
+"metadata_names"
+